split the Summary/Receipt component from the CheckoutSuccess route
