+++ src/components/Summary/Summary.module.css
@@ -1,0 +1,83 @@
+.receipt {
+  padding: 2rem 1rem;
+  max-width: 1200px;
+  margin-left: auto;
+  margin-right: auto;
+}
+
+.infoHeader {
+  border-bottom: solid 1px var(--color-mid-gray);
+  margin-bottom: 1rem;
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  cursor: pointer;
+  h2 {
+    color: var(--color-dark-gray);
+    text-transform: uppercase;
+    font-size: 1rem;
+    font-weight: 400;
+  }
+}
+
+.summaryContainer {
+  margin-bottom: 2rem;
+  display: flex;
+  flex-direction: column;
+  gap: 1rem;
+}
+
+.productCard {
+  display: flex;
+  gap: 0.5rem;
+  background-color: var(--color-white);
+  height: 110px;
+}
+
+.imageWrapper {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  img {
+    width: 100px;
+    object-fit: cover;
+    height: 100%;
+  }
+}
+
+.infoWrapper {
+  padding: 1rem;
+  width: 100%;
+  display: flex;
+  justify-content: space-between;
+  flex-direction: column;
+}
+
+.productTitle {
+  text-decoration: none;
+  h3 {
+    font-size: 1.125rem;
+    font-weight: 600;
+    color: var(--color-black);
+  }
+}
+
+.priceRatingQuantity {
+  display: flex;
+  justify-content: space-between;
+}
+
+.priceRating {
+  display: flex;
+  flex-direction: column;
+  justify-content: space-around;
+  width: 100%;
+  height: 100%;
+}
+
+.quantityWrapper {
+  justify-self: end;
+  align-self: flex-end;
+  font-size: 0.8rem;
+  white-space: nowrap;
+}
